Issue #842 PR feedback: Better name for component and make clickable
